+++ class_items.py
@@ -463,172 +463,6 @@ class Key:
             self.game, f'{who_is_using.name} не знает, как можно использовать ключ если не открывать им что-то.')
 
 
-class Potion:
-    
-    def __init__(self, game, name='Зелье', effect=0, potion_type=0, can_use_in_fight=True):
-        self.game = game
-        self.name = name
-        self.effect = effect
-        self.type = potion_type
-        self.can_use_in_fight = can_use_in_fight
-        self.empty = False
-
-    
-    def __str__(self):
-        return self.description
-
-    
-    def check_name(self, message:str) -> bool:
-        names_list = self.get_names_list(['nom', 'accus'])
-        return message.lower() in names_list
-    
-    
-    def get_names_list(self, cases:list=None) -> list:
-        names_list = ['зелье', 'напиток']
-        for case in cases:
-            names_list.append(self.lexemes.get(case, '').lower())
-        return names_list
-    
-    
-    def on_create(self):
-        return True
-
-    
-    def show(self):
-        return self.description
-
-    
-    def place(self, castle, room=None):
-        if not room:
-            rooms = castle.plan
-            room = randomitem(rooms, False)
-        if room.furniture:
-            furniture = randomitem(room.furniture, False)
-            furniture.put(self)
-        else:
-            room.loot.add(self)
-        return True
-
-    
-    def check_if_can_be_used(self, who_using, in_action: bool) -> bool:
-        game = self.game
-        if self.type == 8 and not who_using.poisoned and who_using.fear == 0:
-            tprint(
-                game, f'{who_using.name} не чувствует никакого недомогания и решает приберечь зелье на попозже.')
-            return False
-        if not in_action and self.type in [0, 3, 5, 7]:
-            tprint(game, 'Это зелье можно использовать только в бою!')
-            return False
-        if in_action and self.type in [1, 2, 4, 6]:
-            tprint(game, 'Это зелье нельзя использовать в бою!')
-            return False
-        return True
-
-    
-    def use_type_0(self, who_using) -> bool:
-        if (who_using.start_health - who_using.health) < self.effect:
-            heal = dice(1, (who_using.start_health - who_using.health))
-        else:
-            heal = dice(1, self.effect)
-        who_using.health += heal
-        text = f'{who_using.name} восполняет {howmany(heal, "единицу жизни,единицы жизни,единиц жизни")}'
-        if who_using.poisoned:
-            who_using.poisoned = False
-            text += ' и излечивается от отравления'
-        tprint(self.game, text)
-        return True
-
-    
-    def use_type_1(self, who_using) -> bool:
-        who_using.start_health += self.effect
-        who_using.health += self.effect
-        tprint(
-            self.game, f'{who_using.name} увеличивает свое максимальное здоровье на {str(self.effect)} до {str(who_using.health)}.')
-        return True
-
-    
-    def use_type_2(self, who_using) -> bool:
-        who_using.stren += self.effect
-        who_using.start_stren += self.effect
-        tprint(
-            self.game, f'{who_using.name} увеличивает свою силу на {str(self.effect)} до {str(who_using.stren)}.')
-        return True
-
-    
-    def use_type_3(self, who_using) -> bool:
-        who_using.stren += self.effect
-        tprint(self.game, f'На время боя {who_using.name} увеличивает \
-            свою силу на {str(self.effect)} до {str(who_using.stren)}.')
-        return True
-
-    
-    def use_type_4(self, who_using) -> bool:
-        who_using.dext += self.effect
-        who_using.start_dext += self.effect
-        tprint(
-            self.game, f'{who_using.name} увеличивает свою ловкость на {str(self.effect)} до {str(who_using.dext)}.')
-        return True
-
-    
-    def use_type_5(self, who_using) -> bool:
-        who_using.dext += self.effect
-        tprint(
-            self.game, f'На время боя {who_using.name} увеличивает свою ловкость на {str(self.effect)} до {str(who_using.dext)}.')
-        return True
-
-    
-    def use_type_6(self, who_using) -> bool:
-        who_using.intel += self.effect
-        who_using.start_intel += self.effect
-        tprint(self.game, f'{who_using.name} увеличивает свой \
-            интеллект на {str(self.effect)} до {str(who_using.intel)}.')
-        return True
-
-    
-    def use_type_7(self, who_using) -> bool:
-        who_using.intel += self.effect
-        tprint(
-            self.game, f'На время боя {who_using.name} увеличивает свой интеллект на {str(self.effect)} до {str(who_using.intel)}.')
-        return True
-
-    
-    def use_type_8(self, who_using) -> bool:
-        who_using.poisoned = False
-        who_using.fear = 0
-        tprint(
-            self.game, f'{who_using.name} излечивается от отравления, избавляется от всех страхов и теперь прекрасно себя чувствует.')
-        return True
-
-    
-    def use(self, who_using, in_action:bool=False):
-        game = self.game
-        functions_list = {
-            0: self.use_type_0,
-            1: self.use_type_1,
-            2: self.use_type_2,
-            3: self.use_type_3,
-            4: self.use_type_4,
-            5: self.use_type_5,
-            6: self.use_type_6,
-            7: self.use_type_7,
-            8: self.use_type_8
-        }
-        if not self.check_if_can_be_used(who_using=who_using, in_action=in_action):
-            return False
-        result = functions_list[self.type](who_using=who_using)
-        if result:
-            who_using.backpack.remove(self)
-        return result
-
-    
-    def take(self, who):
-        if not who.backpack.no_backpack:
-            who.backpack.append(self)
-            tprint(self.game, f'{who.name} забирает {self.lexemes['accus']} себе.')
-            return True
-        return False
-
-
 class Book:
     
     def __init__(self, game, name=None):
